[rollback] Networking
No sure why molesting the DHCP configuration is needed, but I'm sure
that I lost some hours because of it: this should be tested after a
reboot, and you will happily learn that, instead of fixing it, you will
loose the connection to the VM. Do you know who didn't set a local admin
password offering the ability to fix stuff in the OpenStack console? Us.